api/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]